+++ CHANGELOG.md
@@ -4,6 +4,31 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<<<<<<< Updated upstream
+=======
+## [0.6.4] - 2026-01-11
+### Changed
+- **BBMM Buttons:**
+  - Changed "BBMM" buttons on settings and module management windows to display a drop-down menu instead of opening a separate window. 
+### Fixed
+- **Inclusions/Exclusions:**
+  - Fixed bug in inclusions/exclusions migration to persistent storage where when opening a new world, or world that has not migrated could overwrite inclusions/exclusions.
+
+## [0.6.3] - 2026-01-05
+### Fixed
+- **Journal:**
+  - Uploaded image for journal for preset migration instructions. 
+
+## [0.6.2] - 2026-01-05
+### Changed
+- **Changelog:** 
+  - Updated version tag in changelog to reflect the version that it was updated from. Example: "1.0.0 -> 1.2.3". Thanks [ChasarooniZ](https://github.com/ChasarooniZ/)!
+- **Journal:**
+  - Fixed broken image link in journal for preset migration instructions. 
+- **Localization:**
+  - Updated Italian localization. Credit: [GregoryWarn](https://github.com/thejoester/bbmm/issues?q=is%3Apr+is%3Aopen+author%3AGregoryWarn)!
+
+>>>>>>> Stashed changes
 ## [0.6.1] - 2026-01-02
 ### Added
 - **Setting Sync:**